# huff0: Faster 4x decompression